// Client/src/app/Models/Requisition.ts
export class Requisition {
    constructor() {
        this.VehicleTypeID = 0;
    }
    RequisitionDate: Date;
    EmployeeId: number;
    EmployeeName: string;
    PalcetoVisit: string;
    PurposeofVisit: string;
    FromTime: Date;
    ToTime: Date;
    ProjectName: string;
    RequisitionId: number;
    VehicleTypeID: number;
    DriverID: number;
    IsSelfDrive: number;
    SelefDriverName: string;
    NumberofPassenger: number;
    VehicleID: number;
    ApprovedBy: string;
    ApprovedDate: Date;
    RequisitionStatus: number;
    CretedBy: string;
    CreatedDate: Date;
    Status: number;
    StatusName: string;
    Designation:string;
    Department:string;
    VehicleTypeName:string;
    Note:string;
    VehicleName:string;
    ApproveOrRejectNote:string;

}